+++ packages/@tinacms/search/jest.config.js
@@ -2,8 +2,9 @@ import jestRunnerConfig from '@tinacms/scripts/dist/jest-runner.js';
 
 export default {
   ...jestRunnerConfig,
+  extensionsToTreatAsEsm: ['.ts'],
   transform: {
-    '^.+.tsx?$': [
+    '^.+\\.tsx?$': [
       'ts-jest',
       {
         useESM: true,