+++ packages/tinacms-gitprovider-github/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tinacms-gitprovider-github
 
+## 4.1.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@2.0.21
+
+## 4.1.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@2.0.20
+
 ## 4.1.6
 
 ### Patch Changes